Move module script into index.js
